Seguir intentando x64

+++ CustomFaster_MP_TPU/_utils.py
@@ -58,8 +58,8 @@ class BalancedPositiveNegativeSampler:
             num_neg = min(negative.numel(), num_neg)
 
             # randomly select positive and negative examples
-            perm1 = torch.randperm(positive.numel(), device=xm.xla_device())[:num_pos]
-            perm2 = torch.randperm(negative.numel(), device=xm.xla_device())[:num_neg]
+            perm1 = torch.randperm(positive.numel(), device=xm.xla_device(), dtype=torch.int32)[:num_pos]
+            perm2 = torch.randperm(negative.numel(), device=xm.xla_device(), dtype=torch.int32)[:num_neg]
 
             pos_idx_per_image = positive[perm1]
             neg_idx_per_image = negative[perm2]
